Beginning of a GDP-to-MIP transformation base class
